Initial commit for the v3 generator.
This updates the generator to use the new pipeline model introduced in
the new storage SDK.  It works for the storage swagger but is not yet
ready to be used for ARM as there is missing and/or broken codegen when
compared to the v2 generator.
A handful of tests have been updated but a lot have been commented out.