Pageable operations rewrite.
Methods that are pageable will now return a custom page type that is
used to iterate over the pages of data; this "pageable type" utilizes
the same methods as the current implementation but the methods are no
longer exported.
The "*Complete" method has been renamed to "*All" to align with the
Python SDK; it returns an iterator type that builds on top of the
pageable type to seamlessly iterate over the result set, crossing page
boundaries as required.
For long-running operations that return pageable results, the LRO will
return a future that will return a pageable type as its result once the
LRO has completed.