Fix object type codegen to return canonical representation.
In Go a JSON object should be modeled as map[string]interface{}.
Unfortunately this uncovers a broken edge-case in the modeler as it
models objects as a dictionary of objects so you end up with the
following incorrect codegen.
map[string]map[string]interface{}
To work around this I've added a new type InterfaceTypeGo which
represents the empty interface. For dictionaries where the value type
is object return an InterfaceTypeGo object.