Fix method group name collisions in batch generation. (#89)
When AutoRest is run in batch mode the generators aren't recycled per
batch, so as a result any static data persists between batches; this
causes erroneous method group name collisions because the names are
indexed in an object-static hash set.  To fix this include the API
version in the set to ensure they're unique.