Fixed codegen for future's Result() method
If the future.PollingURL() is not the empty string then use that to
retrieve the final result; this is typical for LROs that return both
Location and AsyncOperation headers.
Update validation error to use the new constructor method; this required
updating to the latest go-autorest version.
Fixed some broken tests (port of fix that's in master).