Simplify GetClientPropertyName() (#165)
Use ClientProperty.Name instead of trying to calculate it.  The
calculation is broken for cases where names are camel-cased due to
invalid characters that were removed during naming.  I.e. user.name is
converted to UserName during initial naming; this will translate to a
parameter name of username which incorrectly camel-cases to Username
using the old algorithm.