Remove url.URL param from helper methods
The base URL is on the client so it's no longer required as a parameter.
Verify that the parsed base URL has a scheme.
Skip URL-encoding of path params in presense of x-ms-skip-url-encoding.
Build the complete URL by calling the Parse() method on the base URL
and passing in the relative path.