Add support for additional properties
Add an 'AdditionalProperties' map on applicable types along with a
custom marshaller and unmarshaller.
If a type supports additional properties and also has a field named
'AdditionalProperties' rename the field by appending a '1'.
Note that the current implementation doesn't work for discriminated
types that contain additional properties, or types that use XML.  Will
be fixed in later PRs.