Fix naming
The text helpers in the codegen are not case-preserving and end up
losing information.  Switched to a custom implementation in
ensureNameCase() that uses the algorithm from track 1.
Fixed removePrefix() to perform a case-insensitive comparison.
Removed calls to getEscapedReservedName() where thre is no chance of
collision.
Switch TLA fix-up to use regular expressions.
Use [un]capitalize() functions when only changing the first letter is
required.
Complete the change to treat URIs as strings.
Cleaned up import statements.