Fix marshallers/unmarshallers for MI
Previously the algorithm for determining marshallers and unmarshallers
assumed that a type had only one parent which is incorrect.
Removed all static parent checks, and instead, iterate over all parents.
Changed the signatures of the internal marshallers to accept the raw
object map so that multiple parents can add data to it.
Fixed LROs that return discriminated types.