Remove Connection type
This brings all of the constructors into alignment.
Data-plane client constructors now simply take any host or other client
parameters and a pipeline parameter.
Vanilla ARM client constructors are similar to data-plane, however they
take an azcore.ClientOptions parameter instead of a pipeline.
The parameter logic was ported from the now-deleted connection.ts file
with some slight modifications to simplify it.
Emit a constant for swaggers that specify a host.
Endpoint was renamed to host in places where it made sense.
The client's pipeline field is accessed directly now.