Fix codegen failure for obscure paging corner-case
Predicate using the next page operation on the existence of
nextLinkOperation instead of member as it might be removed when the
underlying operation has been removed via a transform.
Throw an Error if stuttering type name fix-up causes a collision.
Throw an Error if the fallback operation group name causes a collision.
Replaced some deprecated methods with their corresponding updates.