Poller API refactor
The previous design was not mockable due to mixing of exported fields and
methods.  With this new design, the Begin* methods return the poller
directly which contains only methods so they can be mocked.  Some of the
methods have also slightly changed in behavior.
PollUntilDone and Resume methods have moved from the poller response
envelope to the poller, and the poller response envelope is gone.
Poll and Resume now return the response if the LRO reached a successful,
terminal state.
FinalResponse has been removed, its behavior merged into the methods
that can return the final response.