Split options and response type emitter
Models, options, and response types all share the same emitter.  Nice in
theory, but there are enough subtle differences that it's become a bit
of spaghetti code.
Options and response types now have their own emitters.  The emitter for
models will be cleaned up in a separate change.
Added doc comment for the embedded result type in a response envelope.