Decouple code generation stage from M4
Removed code generator's dependency on M4, replacing it with an input
agnostic abstraction layer (gocodemodel.ts).
The naming and transformation phase still interface with M4.
At the end of transformation, the transformed M4 code model is converted
into a GoCodeModel which is then sent to the code generator.
Updated applicable references from gotest and in one case copied the old
code over which still talks to M4.