Split M4 to GoCodeModel into its own phase (#1010)
This reverts the transformation phase to returning an M4 code model so
that test generation isn't broken.
Renamed some of the pipeline phases to more accurately describe what
they do.
Fixed an issue during transformation when creating an InterfaceType and
added armmachinelearning to the test suite which covers it.
Updated baseline files for gotest.