Feature/gocodemodel (#1015)
* Decouple code generation stage from M4 (#1002)

* Decouple code generation stage from M4

Removed code generator's dependency on M4, replacing it with an input
agnostic abstraction layer (gocodemodel.ts).
The naming and transformation phase still interface with M4.
At the end of transformation, the transformed M4 code model is converted
into a GoCodeModel which is then sent to the code generator.
Updated applicable references from gotest and in one case copied the old
code over which still talks to M4.

* propagate module and moduleVersion from session info

ensure that module and containing-module are mutually exclusive

* fix lint

* get azure-arm setting from session

* improve delimited parameters

* pass when there are no unit tests, clean up more lint

* remove duplicated code

* move comment to ctor

* minor cleanup

* Split M4 to GoCodeModel into its own phase (#1010)

This reverts the transformation phase to returning an M4 code model so
that test generation isn't broken.
Renamed some of the pipeline phases to more accurately describe what
they do.
Fixed an issue during transformation when creating an InterfaceType and
added armmachinelearning to the test suite which covers it.
Updated baseline files for gotest.

* Clean up imports in gocodemodel (#1011)

Use a named import for all M4 and GoCodeModel content.  This makes it
clear from where a type originates and simplifies the import statements.

* No more optional arrays (#1013)

Just check for a length of zero where applicable.

* Fix various regressions introduced in GoCodeModel (#1014)

Include polymorphic concrete types when searching for a schema response.
Added support for M4 Credential and ODataQuery types. This brings the
adapter on par with the types supported in the transformer.
Fixed incorrect key name generation for time types.