Get basic codegen for TypeSpec working (#1085)
* Get basic codegen for TypeSpec working

This should cover basic CRUD operations and supporting types. It does
not include support for LROs, pagination, and polymorphism.
The layout and flow is similar to the autorest version; the input is
adapted to the Go code model, fed to the code generator, and the results
written to disk.
There are chunks of commented out code and TODO comments where shortcuts
were taken (e.g. hard-coded values) to get things standing up.
Added "rush tspcompile" which is analogous to "rush regenerate" to
rebuild the world. The only difference is that the script will also mod
tidy after formatting.
The short names for rush script parameters have been removed as they
kept colliding with in-box machinery.
The TS CI has been updated to build/lint/run/publish results.
Centralized initialization of field arrays in the code model.
Added code model support for raw JSON as bytes and slice elements by val
configuration options.
Removed hard-coded pointer to time types in operation request/response
handlers; it now honors the user-specified setting.

* add missing terminal quote

enable verbose flag for tspcompile

* remove mod tidy from tspcompile

this is because it writes to stderr which will cause CI to fail.
added a mod tidy step to CI like the autorest version.

* consolidate shared naming functionality

* add missing license headers