Add coverage for tsp durations
tsp provisions for different encodings besides ISO8601 including numeric
values. This required splitting getPossibleTypes into two functions as
the built-in types are a subset of SdkType.
For header/query collection parameters, set the element type by value to
true as that's what we do in the autorest version.
Add ISO\d+ to the list of acronyms to upper-case.