Add support for multipart/form data in tsp
In tsp, multipart/form params can be model types (M4 promotes the fields
to discrete parameters). Added a model annotation multipartFormData
indicating the model is used to generate multipart/form content. Models
with this annotation omit standard SerDe methods in favor of a helper to
convert the contents to multipart/form data.
Replaced an error comparison for io.EOF with errors.Is.
Added --serve switch to cadl-ranch rush command to start in interactive
model which is useful for debugging failures.