Update to latest tcgc (#1181)
* Update to latest tcgc

The arm flag is now on the sdkContext.
Client initialization is now mandatory and contains an access field
that, when public, means the client is instantiable.
The naming of child clients in hierarchical clients has changed. This is
described in the code comments.
The modeling of host parameters has changed and includes host template
parameters as part of the templateArguments property.
HTTP response status codes can now include a status code range. In
addition, the responses are now in a map.
Various nameInClient properties have been renamed to name.
There is now only one body param for an operation.
The possible api-versions are now modeled as an enum which we skip.
The generatedName field is now a boolean indicating if the value of name
was generated.

* add stutter option for custom stuttering cleanup

* remove workaround that's been fixed