Support tsp spread params
tsp supports spreading a model's fields into discrete params. The values
are bundled back into a single model which is then sent over the wire.
This requires enumerating the tcgc method params, not the operation
params, which correctly models this.
Added type PartialBodyParameter to the code model which represents a
param that's part of a model that's been spread. The resultant model
that contains the spread params is created inline as it doesn't need to
be part of public surface area.