Don't replace error types with azcore.ResponseError
Instances of azcore.ResponseError are returned when an operation fails
and shouldn't be embedded in modeled error types.
Removed omission of foundation error types. It wasn't a great heuristic
and has been superseded by the remove-unreferenced-types option.
Include client ctor params when scanning for unreferenced types.
Make remove-unreferenced-types the default for regeneration script.