Update to the latest TCGC and tsp dependencies (#1385)
* Update to the latest TCGC and tsp dependencies

This includes a breaking change in TCGC how nullable types are
represented. A new kind 'nullable' was introduced that wraps the
underlying type to describe if the type is nullable. Added this kind to
the applicable switch statements, unwrapping as required.
Update emitter version for pending release.

* update autorest.gotest baseline files

in response to 971296885143b12375772775892436668137e142