Remove wire format from model types
The wire format is a function of the operation and isn't tied to a
model. It's also possible for a model to be used with more than one wire
format, although that's not a real scenario at present.
When determining which SerDe methods to generate, find the operation
that uses the model type to determine the wire format.