Emit fake server interceptor
Each fake server has an internal interface var that, when set, can be
used to intercept fake requests. When the return value is true, the fake
request has been "intercepted" and the *http.Response, error values
returned from the interceptor will be used, bypassing dispatching the
request to the fake.