Fix monomorphic response field name (#1517)
* Fix monomorphic response field name

The hard-coded name Value introduced some breaking changes for services
migrating to tsp. In addition, as some tsp-based SDKs have shipped with
this name, we cannot simply revert to the autorest.go behavior.

The recursiveTypeName implementation is essentially a hybrid of the
autorest.go and typespec-go implementations. See the function's doc
comments for an explanation of the naming algorithm. This result is
based on an analysis of released SDKs and a pending release which is
inadvertently being held up due to the result field being renamed.

* update tests

* update release date