Refactor code model type names
Dropped the suffix "Type" from type names as it was redundant and caused
some ugly dependent type names (e.g. ConstantTypeTypes).
BytesType was renamed to EncodedBytes.
LiteralValue was renamed to Literal.
PolymorphicType was renamed to PolymorphicModel.
PrimitiveType was renamed to Scalar.
Fixed sorting of union types.
Added some doc comments. The rest will be added in a subsequent change.