Move moduleName and moduleVersion constants to version.go
This file is emitted for all SDK types. Previously, these constants were
omitted for data-plane with the expectation that they'd be hand-written.
The module-version switch is no longer required when the module switch
is specified. And if module-version is omitted, it defaults to 0.1.0.
This is prep work for emitting client constructors for all SDKs.