Refactor method types in code model (#1712)
* Refactor method types in code model

Added definition for vanilla Go methods and their parts (parameters and
receivers).  The HTTP methods and parameters were updated to extend
these types.  As a result, the client field was renamed to receiver and
the responseEnvelope file was renamed to returns.
Renamed synchronous operation method to SyncMethod so it doesn't clash
with the vanilla Method type.
Removed Client parameter from some codegen methods as it's not required
(the method type contains this info).
Consolidated emitting a method's receiver signature to a shared helper.

* refactor some helpers and consolidate some duplicate code

* improve doc comment