Always prefer client.tsp when available
Fixed the tspcompile script to prefer client.tsp when no tsp file is
specified, and fall back to main.tsp if there is no client.tsp file.
This uncovered a few bugs.
Skip generating fake servers in the fake factory when the client methods
are all internal.
When treating ARM clients as instantiable, set the bit in the code model
that treats the client as instantiable instead of pretending it is (this
consolidates the code paths and avoids duplication).
Model ARM client constructors instead of hard-coding them.
Omit "Method" suffix for an internal method named "internal" when it's a
pager or poller.
Skip creating examples for internal methods.