Refactor handling of tsp endpoint template args
Replace endpoint template values inside client constructors instead of
per method. This means that the replacement only happens once and
removes the need for the template params to be stored on the client.
Refactored the code model to group data unique to instantible clients
into their own type.
The handling of templatedHost, which is unique to Autorest, was also
split into its own type.