Include source package for code model types (#1802)
* Include source package for code model types

This includes clients, client options, constructors, parameter groups,
constants, interfaces, models, and polymorphic models.  For response
envelopes, this can be accessed via its associated method receiver.
Updated getTypeDeclaration helper to take a scope parameter, which is
the package where an instance of a type is being defined.  If the scope
doesn't match the type's package, the package prefix is included.
The constructor for ImportManager now requires a package parameter which
indicates the package where import statements are to be emitted.
Added method addForType to ImportManager which will add the necessary
package import for a type when the type's package doesn't match that of
the ImportManager.
Updated the codegen layer to use getTypeDeclaration instead of making
assumptions about when a package prefix is required.  This consolidates
various bits of code that were adding a package prefix.
Cleaned up the ordering of function parameters in the autorest emitter
to follow a consistent pattern.
Added various function/method doc comments.

* spelling